Add validation for coa type cash & bank

+++ src/views/accounting/master/CoaType.vue
@@ -355,6 +355,10 @@ export default {
           'Hapus Data?',
           'Apakah anda yakin ingin menghapus data ini?')
       ) {
+        if (item.id === 2) {
+          this.$store.dispatch('app/showInfo', 'Tipe akun kas & bank tidak dapat dihapus.')
+          return
+        }
         api.delete(this.endpoint.accounting.coaType, item.id)
           .then(response => {
             if (response.data.success) {
@@ -375,6 +379,10 @@ export default {
         const resp = await api.create(this.endpoint.accounting.coaType, this.data)
         result = resp.data
       } else if (this.data.action === 'edit') {
+        if (this.data.id === 2) {
+          this.$store.dispatch('app/showInfo', 'Tipe akun kas & bank tidak dapat diubah.')
+          return
+        }
         const resp = await api.update(this.endpoint.accounting.coaType, this.data.id, this.data)
         result = resp.data
       }
